+++ src/RessurectIT.Pdf.Print/Dto/Enum/PrintOrientation.cs
@@ -1,0 +1,17 @@
+﻿namespace RessurectIT.Pdf.Print.Dto.Enum;
+
+/// <summary>
+/// Available types of orientations for print options
+/// </summary>
+public enum PrintOrientation
+{
+    /// <summary>
+    /// Orientation of printed page is portrait
+    /// </summary>
+    Portrait,
+
+    /// <summary>
+    /// Orientation of printed page is landscape
+    /// </summary>
+    Landscape
+}