+++ src/RessurectIT.Pdf.Print/Dto/Enum/PrintPaperSize.cs
@@ -1,0 +1,52 @@
+﻿namespace RessurectIT.Pdf.Print.Dto.Enum;
+
+/// <summary>
+/// Available types of paper sizes for print options
+/// </summary>
+public enum PrintPaperSize
+{
+    /// <summary>
+    /// Page size A2
+    /// </summary>
+    A2,
+
+    /// <summary>
+    /// Page size A3
+    /// </summary>
+    A3,
+
+    /// <summary>
+    /// Page size A4
+    /// </summary>
+    A4,
+
+    /// <summary>
+    /// Page size A5
+    /// </summary>
+    A5,
+
+    /// <summary>
+    /// Page size A6
+    /// </summary>
+    A6,
+    
+    /// <summary>
+    /// Page size letter
+    /// </summary>
+    Letter,
+
+    /// <summary>
+    /// Page size legal
+    /// </summary>
+    Legal,
+
+    /// <summary>
+    /// Page size tabloid
+    /// </summary>
+    Tabloid,
+
+    /// <summary>
+    /// Page size statement
+    /// </summary>
+    Statement
+}