+++ src/RessurectIT.Pdf.Print/Dto/Enum/PrintScale.cs
@@ -1,0 +1,22 @@
+﻿namespace RessurectIT.Pdf.Print.Dto.Enum;
+
+/// <summary>
+/// Available types of scale for print options
+/// </summary>
+public enum PrintScale
+{
+    /// <summary>
+    /// No scaling will be performed
+    /// </summary>
+    Noscale, 
+
+    /// <summary>
+    /// Page will be shrink
+    /// </summary>
+    Shrink,
+
+    /// <summary>
+    /// Page will fit
+    /// </summary>
+    Fit
+}