+++ src/RessurectIT.Pdf.Print/Dto/Enum/PrintSide.cs
@@ -1,0 +1,27 @@
+﻿namespace RessurectIT.Pdf.Print.Dto.Enum;
+
+/// <summary>
+/// Available types of sides for print options
+/// </summary>
+public enum PrintSide
+{
+    /// <summary>
+    /// Duplex printing
+    /// </summary>
+    Duplex,
+
+    /// <summary>
+    /// Duplex short printing
+    /// </summary>
+    Duplexshort,
+
+    /// <summary>
+    /// Duplex long printing
+    /// </summary>
+    Duplexlong,
+
+    /// <summary>
+    /// Simplex printing
+    /// </summary>
+    Simplex
+}