+++ src/RessurectIT.Pdf.Print/Dto/Enum/PrintSubset.cs
@@ -1,0 +1,17 @@
+﻿namespace RessurectIT.Pdf.Print.Dto.Enum;
+
+/// <summary>
+/// Available types of subsets for print options
+/// </summary>
+public enum PrintSubset
+{
+    /// <summary>
+    /// Prints only odd pages
+    /// </summary>
+    Odd,
+
+    /// <summary>
+    /// Prints only even pages
+    /// </summary>
+    Even
+}