+++ src/RessurectIT.Pdf.Print/Dto/PrintOptions.cs
@@ -1,4 +1,6 @@
-﻿namespace RessurectIT.Pdf.Print.Dto;
+﻿using RessurectIT.Pdf.Print.Dto.Enum;
+
+namespace RessurectIT.Pdf.Print.Dto;
 
 /// <summary>
 /// Request storing information about what should be printed
@@ -7,17 +9,113 @@ public class PrintOptions
 {
     #region public properties
 
-    //string printer = 2;
-    //string pages = 3;
-    //string subset = 4;
-    //string orientation = 5;
-    //string scale = 6;
-    //bool monochrome = 7;
-    //string side = 8;
-    //string bin = 9;
-    //string paperSize = 10;
-    //bool silent = 11;
-    //bool printDialog = 12;
-    //int32 copies = 13;
+    /// <summary>
+    /// Send a file to the specified printer. Id of printer.
+    /// </summary>
+    public string? Printer
+    {
+        get;
+        set;
+    }
+
+    //TODO: regex validator
+    /// <summary>
+    /// Specifies which pages to print in the PDF document. Supports ranges (1-5), supports comma separated pages (3,5) and combinations (1-3,5)
+    /// </summary>
+    public string? Pages
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Will print odd pages only when value is odd. Will print even pages only when even.
+    /// </summary>
+    public PrintSubset? Subset
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Can provide 90-degree rotation of contents (NOT the rotation of paper which must be pre-set by the choice of printer defaults).
+    /// </summary>
+    public PrintOrientation? Orientation
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Supported names noscale, shrink and fit.
+    /// </summary>
+    public PrintScale? Scale
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Prints the document in black and white. Default is false.
+    /// </summary>
+    public bool? Monochrome
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Supported names duplex, duplexshort, duplexlong and simplex.
+    /// </summary>
+    public PrintSide? Side
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Select tray to print to. Number or name.
+    /// </summary>
+    public string? Bin
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Specifies the paper size. Supported names A2, A3, A4, A5, A6, letter, legal, tabloid, statement.
+    /// </summary>
+    public PrintPaperSize? PaperSize
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Silences SumatraPDF's error messages.
+    /// </summary>
+    public bool? Silent
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Displays the Print dialog for all the files indicated on this command line.
+    /// </summary>
+    public bool? PrintDialog
+    {
+        get;
+        set;
+    }
+
+    /// <summary>
+    /// Specifies how many copies will be printed.
+    /// </summary>
+    public int? Copies
+    {
+        get;
+        set;
+    }
     #endregion
 }